requirements.txt
aniso8601==8.0.0
attrs==19.3.0
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
-e git+git@github.com:phaesoo/flask-jwt-auth.git@a9fc19f9e58c86c312cc90517b1c649702c93ce7#egg=flask_jwt_auth_example
flask-restplus==0.13.0
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
pathlib2==2.3.5
pluggy==0.13.1
py==1.8.1
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.15.6
pytest==5.3.2
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
wcwidth==0.1.8
Werkzeug==0.16.0
zipp==0.6.0
